Support for LLVM IR funnel shift left/right intrinsic
LangRef: https://releases.llvm.org/7.0.0/docs/LangRef.html#llvm-fshl-intrinsic

```
The ‘llvm.fshl’/‘llvm.fshr’ family of intrinsic functions performs
a funnel shift left/right: the first two values are concatenated
as { %a : %b } (%a is the most significant bits of the wide value),
the combined value is shifted left/right,
and the most/least significant bits are extracted to produce a result
that is the same size as the original arguments.
<...>
For vector types, the operation occurs for each element of the vector.
The shift argument is treated as an unsigned amount modulo
the element size of the arguments.
```

Questions:
* Should there be constant-evaluation fastpaths in these `expr::rotate()` / `expr::funnelShift()`?
  I have avoided adding them for now because the bit width where these computations are
  done matters, and right now all such `binop_ufold()` implicitly widen to 64-bit.
  Hmm, but now that i think about it, maybe it would still work, will take a second look..
* Any specific tests i'm missing?
* Should this be a `TernaryOp`, or should i just add `FunnelShiftOp`?

Bugs:
* There seems to be something off with `poison` propagation.
  See `tests/unit/fshl/poison-7.opt` - that **should** be poison,
  we combined a known-poison value with something else.
  It does work without all that extra instructions - see `tests/unit/fshl/poison-7-v2.opt`

Observations: alive2 api has a few rough edges
* `zext` takes the increment, not the destination width
* `extract` takes bit positions instead of `{skip, len}`
* I really don't like non-**explicit** single-argument constructors.
  I have stumbled into a few bugs because of this here already, e.g.:
```c++
  auto nbits = bits(); // decltype(nbits) will be `unsigned`!
  expr rhs_mod_width = rhs.urem(nbits); // compiles, fails at run time
```

Similar to https://github.com/google/souper/pull/406