TransformVerify: only check src instrs that exist in dst too
At least the lhs-intermediate-variable.opt test causes a crash otherwise.
Because we go through every instruction on the LHS, and verify it against
it's counterpart on RHS. But naturally, if said instruction only exists
on one side, we can't get it's counter-part from the other side
so the std::unordered_map<>::at() throws an exception, and alive crashes.