Vector reductions: unary integral reductions
https://llvm.org/docs/LangRef.html#experimental-vector-reduction-intrinsics
> Horizontal reductions of vectors can be expressed using the
> following intrinsics. Each one takes a vector operand as an input
> and applies its respective operation across all elements of the vector,
> returning a single scalar result of the same element type.

Now that since https://reviews.llvm.org/D80867 /
https://github.com/llvm/llvm-project/commit/e50059f6b6b32c134e0f4f81281a774cb460524e
the (experimental) reductions are considered to be somewhat non-outcasts
and are produced by the middle-end, i feel like this is the perfect time
for alive2 to learn about them :)

The reductions can be classified into four groups:
* integral math/logic (this pr) - `add`/`mul` / `and`/`or`/`xor`
* integral aggregationh - `smax`/`smin`/`umax`/`umin`
* fp math - `fadd`/`fmul`
* fp aggregationh - `fmax`/`fmin`

As it can be seen, this patch adds support for the first group only.
Rest to follow.

I have considered not creating new `UnaryReductionOp`,
but using `UnaryOp`, but it ends up being really ugly because
these operations change the type of operand. In particular,
we'd need to spell them something like `reduce_add <2 x i8> %x to i8`,
and the type printing won't roundtrip.

Thoughts?

CC @rotateright, @topperc