run TV from our LLVM plugin in parallel by forking (#603)
* run TV from our LLVM plugin in parallel by forking. provide two
parallelism managers: one that just forks in uncontrolled fashion (for
testing and debugging -- don't run it on large inputs) and another
that uses the jobserver from GNU make to limit parallelism.

when output in not directed to a report directory, it all just gets
blatted to the terminal in mixed-up fashion by all the different
proceses. when output is directed to a report directory, it goes to a
lot of little individual files and then a utility fix_report_files
stitches it all back together.