Fixes to parallel mode (#621)
well, I hadn't originally realized what things like the sqlite 
amalgamation look like, but they contain a pretty huge number of
functions. this patch avoids leaking file descriptors in the top-level
clang process and also avoids all operations that have runtime
proportional to the total number of forked processes (other than the
one for putting together all the output at the end, which is
unavoidable). also now we use poll() to efficiently see if any child
processes have something to say to us, and moreover manage a
persistent pool of pollfd structs, so this does not have to be
constructed on the fly every time.

also: fix a deadlock in the signal handler which was calling into some
iostream code that malloced -- try to be more async-safe

also: test parallel mode on Mac, seems to work fine

also: make parallel output to console work, there was a silly omission