Print output earlier (#635)
* previously, the main clang process could use a lot of RAM since it
buffered output from all children and then dumped it all at once at
the end. this hurt when using alivecc on huge files like sqlite3.c.

this patch dumps output as it goes along. the only tricky thing was
dealing with the fact that children complete out of order.

this also does just a tiny bit of cleanup on the parallel manager.