add option to limit number of bits used for GEP offsets (#661)
This patch adds a max_offset_bits config option, to limit the maximum
value assumed for offsets. This effectively provides an option to
limit the width of pointer arithmetic, which in turn speeds up
verification of programs using a lot of pointer arithmetic.

For regular arithmetic, a common advice is to limit the width of integer
types, but there is no easy way to  do the same thing for pointers. The
new option allows limiting the width of pointer arithmetic globally.

This should not impact soundness & completeness, unless the input
program has pointer arithmetic that does not fit in the specified width
(e.g. has a GEP that uses an offset outside the max range). So the users
of the option should think carefully when using the option.

Perhaps there is also a way to verify that the chosen value is not too
small?