two fixes to fpto*i
As mentioned in https://reviews.llvm.org/D106053#2885236 , Alive2 had a bug in fptoui/si encoding.

`fptoui float 31.5 to i32' was yielding poison because np was `(float)(i32)31.5 == 31.5' which is alway false.
The rhs is fixed to round towards zero so that it becomes `(float)(i32)31.5 == round_towards_zero(31.5)'.

Also, fptoui should round towards zero according to LangRef, but it was using nearest ties to even.
This is fixed as well.