Fix conflicting types error when taking maximum of alignments
When using llvm version llvmorg-13.0.0 and GCC libstdc++ v11.2.1
a deduced conflicting types error occurs when taking the maximum
of IR::align and DataLayout::getABITypeAlignment(...). This
results in arguments of type unsigned long and unsigned int to
std::max. The resolution is to use a decltype of the addition of
the two arguments and perform static casting from that.