Troubleshooting suggestion for Mac dynamic_lookup/ chained fixups linker warning and subsequent “symbol not found in flat namespace” crash. Thanks to Ronald Oussoren (https://github.com/python/cpython/issues/97524#issuecomment-1291003452).
This works and seems needed on my ARM M1 MacOS 12.7 21G816, when compiling with either Apple clang-1400.0.29.202
or Homebrew clang version 16.0.5.  The discussion there suggests that the Clang in Xcode 14.3 might fix this, but that
won’t run on my MacOS.

I’m not submitting my change to the CMake file (https://github.com/FlashSheridan/alive2/commit/7c642778),
since it presumably isn’t affecting most users and might be a performance hit.  Feel free to pull it if you disagree.