ir: thread parent BB and Function throughout
Alive IR suffers from the deficiency that the parent BasicBlock is not
known in the Instr class, and the parent Function is not known in the
BasicBlock class. The motivation for this patch is that, to support
certain attributes like vscale, the dynamic vscale value (which will
only be available on the function) needs to be known in the Instr class
when checking programs with scalable vector types. The side-effect of
this patch is that it cleans up the API of several dup() functions to be
more intuitive.